Simply DynamoidReset

+++ spec/support/dynamoid_reset.rb
@@ -1,41 +1,15 @@
 module DynamoidReset
   def self.create
-    tries = 0
-    until table_deleted? || tries >= 5 do
-      puts("Sleeping for #{tablename} to be deleted. #{tries} tries")
-      sleep(15)
-      tries += 1
-    end
+    self.delete
+    Dynamoid.adapter.tables.clear
 
-    BookIndexing.create_table(sync: true)  #there is no sync: true for delete tho....
-
-    tries = 0
-    until table_created? || tries >= 5 do
-      puts("Sleeping for #{tablename} to be created. #{tries} tries.")
-      sleep(5)
-      tries +=1
-    end
-  end
-
-  def self.table_created?
-    tables = Dynamoid.adapter.list_tables
-    puts "Checking for table_created? found Dynamo tables - '#{tables.join(', ')}'"
-    tables.include? BookIndexing.table_name
-  rescue
-    false
-  end
-
-  def self.table_deleted?
-    tables = Dynamoid.adapter.list_tables
-    puts "Checking for table_deleted? found Dynamo tables - '#{tables.join(', ')}'"
-    tables.exclude? BookIndexing.table_name
-  rescue
-    false
+    BookIndexing.create_table(sync: true)
   end
 
   def self.delete
-    Dynamoid.adapter.delete_table(tablename)
-    Dynamoid.adapter.tables.clear    #clear the table cache
+    if Dynamoid.adapter.list_tables.include?(tablename)
+      Dynamoid.adapter.delete_table(tablename)
+    end
   end
 
   def self.tablename
